NOTICE: I MADE LOT OF CHANGE AND THE NEXT COMMITS RELATED TO IT - Support TCP traceroute - Restructure the codebase - Make things modular - Accurate results based on matching each response to sent probe request - Add feature early stop on silence (consecutive no replies) - Support dealy between sending probe requests - Document codebase
